[ISSUE-1254] Move Ticket APIs to jmap-extensions-api
