+++ tmail-backend/jmap/extensions-api/src/main/scala/com/linagora/tmail/james/jmap/ticket/ForbiddenException.scala
@@ -1,0 +1,3 @@
+package com.linagora.tmail.james.jmap.ticket
+
+case class ForbiddenException() extends RuntimeException
