+++ tmail-backend/jmap/extensions-api/src/main/scala/com/linagora/tmail/james/jmap/ticket/TicketStore.scala
@@ -1,0 +1,11 @@
+package com.linagora.tmail.james.jmap.ticket
+
+import reactor.core.scala.publisher.SMono
+
+trait TicketStore {
+  def persist(ticket: Ticket): SMono[Unit]
+
+  def retrieve(value: TicketValue): SMono[Ticket]
+
+  def delete(ticketValue: TicketValue): SMono[Unit]
+}
